refactor: Extract MixinDate.dateToString()
Instead of using toLocaledateString() directly inside the DOM template,
create a helper method `dateToString()` and supply it using a mixin.
To do so, create a new Vue mixin named `MixinDate`, available in
`/js/mixin-date.js`.

Note: This also uses YYYY-MM-DD format regardless of browser or locale.
Note: This also eliminates an incorrect call to toLocaleDateString()
      (the first option should have been `undefined`, not `{}`).
Note: The `dateStyle` option is supported only in Chrome, so the code
      was pretty faulty to begin with. A good thing we eliminated it!