chore: Use vue-ls to indirectly acquire Vue types
Because TypeScript does not pull in `@types/vue` for type checking, we
have to acquire types from another typing package that depends on the
`vue` package. We choose `vue-ls`, even thouh we do not actually use it,
just so that VS Code correctly pulls in the `vue` package for type
checking.